+++ index.js
@@ -6,24 +6,31 @@ var recorder = require('./source/recorder');
 var playback = require('./source/playback');
 var mkdirp = require('mkdirp');
 var colors = require('colors');
+var glob = require('glob');
 
-var DEFAULT_SCREEN_SIZE = [1024, 768];
+// optimal default screen size. 1200 is bootstrap's definition of 'large screen'
+// and 795 is a mba 13inch's available height for firefox window in Selenium.
+// The actual height of the chromeless viewport should be 689
+var DEFAULT_SCREEN_SIZE = [1200, 795];
 // TODO: integration with remote environment
-var currentPath = process.cwd();
 
-function _getTaskFolderName(taskName) {
-  return currentPath + '/' + taskName + '.hux';
-}
+// TODO: following
+// Btw, whenever 'path' is mentioned and it concerns a file, the file's name
+// itself isn't included
 
-// the signature of the operation passed:
-// operation(browserName, singleTaskObj, callback)
-function _operateOnEachTask(browserName, operation) {
+// the signature of the action passed:
+// action(browserName, singleTaskObj, callback)
+// currently one of:
+// `_recordTask`
+// `_playbackTaskAndSaveScreenshot`
+// `_playbackTaskAndCompareScreenshot`
+function _operateOnEachHuxleyfile(browserName, huxleyfilePath, action, next) {
   var tasks;
   try {
-    tasks = require(currentPath + '/Huxleyfile.json');
+    tasks = require(huxleyfilePath + '/' + 'Huxleyfile.json');
   } catch (err) {
-    console.error('No runnable Huxleyfile.json found at %s.'.red, currentPath);
-    return;
+    console.log(err);
+    return next('Failed to load Huxleyfile in ' + huxleyfilePath + '.');
   }
 
   // filter out every task that's marked skipped (i.e. has the key `xname`
@@ -33,64 +40,75 @@ function _operateOnEachTask(browserName, operation) {
   });
 
   if (tasks.length === 0) {
-    return console.error('No runnable task found at %s.'.red, currentPath);
+    return next('No runnable Huxleyfile.json found at ' + huxleyfilePath + '.');
   }
 
   if (unSkippedTasks.length === 0) {
-    console.error('Every task is marked as skipped at %s.'.yellow, currentPath);
-    return;
+    console.warn(
+      '\nEvery task is marked as skipped at %s.\n'.yellow, huxleyfilePath
+    );
+    return next();
   }
 
   var currentTaskCount = 0;
 
-  operation(browserName,
-            unSkippedTasks[currentTaskCount],
-            function runNextTask(err) {
-    if (err) {
-      process.stdin.pause();
-      return console.error(typeof err === 'string' ? err.red : err.message.red);
-    }
+  action(browserName,
+        huxleyfilePath,
+        unSkippedTasks[currentTaskCount],
+        function runNextTask(err) {
+
+    if (err) return next(err);
 
     if (currentTaskCount === unSkippedTasks.length - 1) {
-      process.stdin.pause();
-      console.log('\nAll done successfully!'.green);
       var skippedTaskCount = tasks.length - unSkippedTasks.length;
       if (skippedTaskCount > 0) {
         console.log(
-          '\nYou\'ve skipped %s task%s at %s'.yellow,
+          '\nYou\'ve skipped %s task%s at %s\n'.yellow,
           skippedTaskCount,
           skippedTaskCount > 1 ? 's' : '',
-          currentPath
+          huxleyfilePath
         );
       }
+      next();
     } else {
       console.log('\nNext task...\n');
-      operation(browserName, unSkippedTasks[++currentTaskCount], runNextTask);
+      action(
+        browserName,
+        huxleyfilePath,
+        unSkippedTasks[++currentTaskCount],
+        runNextTask
+      );
     }
   });
 }
 
-function _recordTask(browserName, task, next) {
+function _recordTask(browserName, huxleyfilePath, task, next) {
   var driver = browser.getNewDriver(browserName);
   if (driver == null) return next('Unsupported browser.');
 
   var screenSize = task.screenSize || DEFAULT_SCREEN_SIZE;
 
   browser.openToUrl(driver, task.url, screenSize[0], screenSize[1], function() {
-    console.log('Running test: ' + task.name);
+    console.log('Running test: %s'.bold, task.name);
     recorder.startPromptAndInjectEventsScript(driver,
                                               function(screenShotTimes) {
       recorder.stopAndGetProcessedEvents(driver,
                                         screenShotTimes,
                                         function(allEvents) {
-        _saveTaskAsJsonToFolder(task.name, allEvents, function(err) {
+        _saveTaskAsJsonToFolder(task.name,
+                                huxleyfilePath,
+                                allEvents,
+                                function(err) {
           console.log(
-            '\nDon\'t move! Simulating the recording now...\n'.bold.yellow
+            '\nDon\'t move! Simulating the recording now...\n'.yellow
           );
 
           browser.quit(driver, function() {
             if (err) return next(err);
-            _playbackTaskAndSaveScreenshot(browserName, task, function(err) {
+            _playbackTaskAndSaveScreenshot(browserName,
+                                          huxleyfilePath,
+                                          task,
+                                          function(err) {
               if (err) return next(err);
               next();
             });
@@ -101,9 +119,9 @@ function _recordTask(browserName, task, next) {
   });
 }
 
-function _saveTaskAsJsonToFolder(taskName, taskEvents, next) {
+function _saveTaskAsJsonToFolder(taskName, huxleyfilePath, taskEvents, next) {
   // `taskEvents` should already have been processed by `_processRawTaskEvents`
-  var folderPath = _getTaskFolderName(taskName);
+  var folderPath = huxleyfilePath + '/' + taskName + '.hux';
   mkdirp(folderPath, function(err) {
     if (err) return next(err);
     fs.writeFile(folderPath + '/record.json',
@@ -115,19 +133,31 @@ function _saveTaskAsJsonToFolder(taskName, taskEvents, next) {
   });
 }
 
-function _playbackTaskAndSaveScreenshot(browserName, task, next) {
-  _playbackTask(browserName, task, false, next);
+function _playbackTaskAndSaveScreenshot(browserName,
+                                        huxleyfilePath,
+                                        task,
+                                        next) {
+  _playbackTask(browserName, huxleyfilePath,task, false, next);
 }
 
-function _playbackTaskAndCompareScreenshot(browserName, task, next) {
-  _playbackTask(browserName, task, true, next);
+function _playbackTaskAndCompareScreenshot(browserName,
+                                          huxleyfilePath,
+                                          task,
+                                          next) {
+  _playbackTask(browserName, huxleyfilePath,task, true, next);
 }
 
-function _playbackTask(browserName, task, compareInsteadOfOverride, next) {
+function _playbackTask(browserName,
+                      huxleyfilePath,
+                      task,
+                      compareInsteadOfOverride,
+                      next) {
   var taskEvents;
+  var recordPath = huxleyfilePath + '/' + task.name + '.hux';
+
   try {
-    taskEvents = require(_getTaskFolderName(task.name) + '/record.json');
-  } catch (e) {
+    taskEvents = require(recordPath + '/record.json');
+  } catch (err) {
     return next('Cannot find enough info on recorded actions.');
   }
 
@@ -137,10 +167,10 @@ function _playbackTask(browserName, task, compareInsteadOfOverride, next) {
   var screenSize = task.screenSize || DEFAULT_SCREEN_SIZE;
 
   browser.openToUrl(driver, task.url, screenSize[0], screenSize[1], function() {
-    console.log('Running test: ' + task.name);
+    console.log('\nRunning test: %s\n', task.name);
 
     var options = {
-      taskPath: _getTaskFolderName(task.name),
+      taskPath: recordPath,
       compareWithOld: compareInsteadOfOverride
     };
     playback(driver, taskEvents, options, function(err) {
@@ -152,16 +182,62 @@ function _playbackTask(browserName, task, compareInsteadOfOverride, next) {
   });
 }
 
-function recordTasks(browserName) {
-  _operateOnEachTask(browserName, _recordTask);
+// the path doesn't include the name `Huxleyfile.json`
+function _operateOnAllHuxleyfiles(browserName, huxleyfilePath, action) {
+  huxleyfilePath = huxleyfilePath || '**/';
+  huxleyfilePath = process.cwd() + '/' + huxleyfilePath;
+
+  // use glob to find every huxleyfile in the path, including nested ones.
+  // Normally we'd do a simple exec('find blabla'), but this wouldn't work on
+  // Windows. So search every huxleyfile location, and trim the file name to get
+  // the container folders, needed for storing screenshots and such
+  var allHuxleyPaths = glob.sync(huxleyfilePath + '/Huxleyfile.json');
+  allHuxleyPaths = allHuxleyPaths.map(function(path) {
+    return path.substr(0, path.lastIndexOf('/'));
+  });
+
+  if (allHuxleyPaths.length === 0) {
+    return console.error('No Huxleyfile.json found anywhere.'.red);
+  }
+
+  var currentHuxleyfileIndex = 0;
+  _operateOnEachHuxleyfile(browserName,
+                          allHuxleyPaths[currentHuxleyfileIndex],
+                          action,
+                          function runNextHuxleyfile(err) {
+    if (err) {
+      process.stdin.pause();
+      return console.error(typeof err === 'string' ? err.red : err.message.red);
+    }
+    if (currentHuxleyfileIndex === allHuxleyPaths.length - 1) {
+      process.stdin.pause();
+      console.log('\nAll done successfully!\n'.green);
+    } else {
+      currentHuxleyfileIndex++;
+      _operateOnEachHuxleyfile(
+        browserName,
+        allHuxleyPaths[currentHuxleyfileIndex],
+        action,
+        runNextHuxleyfile
+      );
+    }
+  });
 }
 
-function playbackTasksAndSaveScreenshots(browserName) {
-  _operateOnEachTask(browserName, _playbackTaskAndSaveScreenshot);
+function recordTasks(browserName, huxleyfilePath) {
+  _operateOnAllHuxleyfiles(browserName, huxleyfilePath, _recordTask);
 }
 
-function playbackTasksAndCompareScrenshots(browserName) {
-  _operateOnEachTask(browserName, _playbackTaskAndCompareScreenshot);
+function playbackTasksAndSaveScreenshots(browserName, huxleyfilePath) {
+  _operateOnAllHuxleyfiles(
+    browserName, huxleyfilePath, _playbackTaskAndSaveScreenshot
+  );
+}
+
+function playbackTasksAndCompareScrenshots(browserName, huxleyfilePath) {
+  _operateOnAllHuxleyfiles(
+    browserName, huxleyfilePath, _playbackTaskAndCompareScreenshot
+  );
 }
 
 module.exports = {
@@ -169,6 +245,3 @@ module.exports = {
   playbackTasksAndSaveScreenshots: playbackTasksAndSaveScreenshots,
   playbackTasksAndCompareScrenshots: playbackTasksAndCompareScrenshots,
 };
-
-// recordTasks();
-